+++ src/shared/entities.d.ts
@@ -1,4 +1,4 @@
-import { ProjectVisibility, REGION_LABELS } from "./constants";
+import { ProjectVisibility, ReferenceLayerTypes, REGION_LABELS } from "./constants";
 
 export type UserId = string;
 
@@ -178,6 +178,19 @@ export interface IRegionConfig {
   readonly archived: boolean;
 }
 
+export interface IReferenceLayer {
+  readonly id: ReferenceLayerId;
+  readonly name: string;
+  readonly layer_type: ReferenceLayerTypes.Point | ReferenceLayerTypes.Polygon;
+  readonly label_field: string;
+}
+
+export interface ReferenceLayerProperties {
+  // key is a property name on the geojson
+  // value is the value for that property
+  readonly [id: string]: number | string;
+}
+
 interface ProjectTemplateFields {
   readonly name: string;
   readonly regionConfig: IRegionConfig;
@@ -189,6 +202,8 @@ interface ProjectTemplateFields {
 }
 
 export type ProjectId = string;
+
+export type ReferenceLayerId = string;
 
 export type IProject = ProjectTemplateFields & {
   readonly id: ProjectId;
@@ -225,6 +240,13 @@ export interface CreateProjectData {
   readonly districtsDefinition?: DistrictsDefinition;
   readonly populationDeviation?: number;
   readonly projectTemplate?: Pick<IProjectTemplate, "id">;
+}
+
+export interface CreateReferenceLayerData {
+  readonly name: string;
+  readonly project: Pick<IProject, "id">;
+  readonly layer_type: ReferenceLayerTypes.Point | ReferenceLayerTypes.Polygon;
+  readonly label_field: string;
 }
 
 export type UpdateProjectData = Pick<