+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Add support for uploading reference layers to a project [#902](https://github.com/PublicMapping/districtbuilder/pull/902)
+
 ### Changed
 ### Fixed
 
@@ -31,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add levers to tune LB health check configuration [#950](https://github.com/PublicMapping/districtbuilder/pull/950)
+
 
 ### Changed
 - Parameterize `max_old_space_size` [#942](https://github.com/PublicMapping/districtbuilder/pull/942)